refactor: 프로젝트 상세 페이지 레이아웃 구조 정리 및 리뷰 반영
