--- frontend/src/app/(project)/projects/[projectId]/_component/ProjectDetailLinkItem.tsx
@@ -1,28 +1,0 @@
-interface ProjectDetailLinkItemProps {
-    label: string;
-    href: string;
-    onContextMenu?: (e: React.MouseEvent<HTMLAnchorElement>) => void;
-}
-
-export const ProjectDetailLinkItem = ({
-                                          label,
-                                          href,
-                                          onContextMenu,
-                                      }: ProjectDetailLinkItemProps) => {
-    const faviconUrl = `https://www.google.com/s2/favicons?domain=${href}&sz=64`;
-
-    return (
-        <a
-            href={href}
-            target="_blank"
-            rel="noreferrer"
-            onContextMenu={onContextMenu}
-            className="flex items-center gap-2 rounded-md px-2 py-1 transition-colors hover:bg-fill-normal"
-        >
-            <img src={faviconUrl} alt="favicon" className="h-5 w-5 shrink-0" />
-            <span className="text-caption-1 font-semibold text-label-neutral">
-        {label}
-      </span>
-        </a>
-    );
-};
