+++ frontend/src/app/(project)/projects/[projectId]/page.tsx
@@ -1,26 +1,27 @@
 'use client';
 
-import { useState } from 'react';
+import Link from 'next/link';
 
-import { EXAMPLE_USERS } from '@/constants/exampleConstant';
-import type { UserInfo } from '@/components/commons/UserAvatarGroup';
-import { ProjectDetailHeader } from './_component/ProjectDetailHeader';
-import { ProjectDetailLinks } from './_component/ProjectDetailLinks';
-import { ProjectDetailPageContent } from './_component/ProjectDetailPageContent';
-import type { ProjectViewTypes } from './_component/ProjectViewTabs';
+import { useProjectDetailLayout, type ProjectViewTypes } from '@/app/(project)/projects/[projectId]/layout';
+
+const VIEW_LABELS: Record<ProjectViewTypes, string> = {
+  'node-flow': '노드 플로우',
+  list: '리스트',
+  kanban: '칸반',
+};
 
 export default function ProjectDetailPage() {
-    const [activeView, setActiveView] = useState<ProjectViewTypes>('node-flow');
+  const { activeView } = useProjectDetailLayout();
 
-    return (
-        <div className="flex h-full w-full flex-1 flex-col overflow-hidden">
-            <ProjectDetailHeader
-                activeView={activeView}
-                onlineUsers={EXAMPLE_USERS as readonly UserInfo[]}
-                onViewChange={setActiveView}
-            />
-            <ProjectDetailLinks />
-            <ProjectDetailPageContent activeView={activeView} />
-        </div>
-    );
+  return (
+    <section className="flex flex-1 flex-col items-center justify-center gap-3 overflow-hidden bg-surface-canvas px-20 py-24">
+      <p className="text-body-1 font-medium text-label-neutral">{VIEW_LABELS[activeView]}</p>
+      <Link
+        href="/projects"
+        className="rounded-md border border-line-normal-neutral bg-static-white px-4 py-2 text-title-3 font-medium text-label-normal hover:bg-fill-alternative"
+      >
+        뒤로가기
+      </Link>
+    </section>
+  );
 }
